splitting SignatureLib tests
